Added new event, allows blank values, indentation
Added a new event, onInit, that fires before the chart renders. Data is
passed so a script can check if the jquery chart will be empty or not.
Blank dates can be used to create a 'section' in the chart, and
indentation css allows some items to be indented for grouping